requirements/base.txt
# Common
celery==5.2.7
dj-database-url==0.5.0
greenlet==1.1.3.post0
# hiredis==2.0.0
logwood==3.1.0
# marshmallow-union==0.1.15.post1
# mip==1.6.2
# pytz==2022.1
psycopg2-binary==2.9.5  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django==4.1.3  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# django-extensions==3.1.5
django-celery-beat==2.4.0
ipython==8.6.0
